refactor localstorage login back to App.js
